Get data as an xlsx file
With this function, we are able to retrieve that unified dataset as an xlsx file, in which each of the original datasets would have its own datasheet. What could be improved upon is that datasheet 1 currently is storing all the input into just 2 rows due to how the JSON data structure was formatted. Have to work on this